@@ -3,7 +3,7 @@ name = "simple_box"
 version = "0.1.0"
 authors = ["Charles Bournhonesque <charlesbour@gmail.com>"]
 edition = "2021"
-rust-version = "1.65"
+rust-version = "1.76"
 description = "Examples for the lightyearServer-client networking library for the Bevy game engine"
 readme = "README.md"
 repository = "https://github.com/cBournhonesque/lightyear"
@@ -13,21 +13,21 @@ license = "MIT OR Apache-2.0"
 publish = false
 
 [features]
+default = ["server", "client"]
+gui = ["lightyear_examples_common/gui", "dep:bevy_mod_picking"]
+client = ["lightyear_examples_common/client", "gui"]
+server = ["lightyear_examples_common/server"]
 metrics = ["lightyear/metrics", "dep:metrics-exporter-prometheus"]
 
 [dependencies]
-lightyear_examples_common = {workspace = true, default-features = true}
-lightyear = {workspace = true, features = [
-  "steam",
-  "webtransport",
-  "websocket",
-] }
+lightyear_examples_common.workspace = true
+lightyear.workspace = true
 serde.workspace = true
 anyhow.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
 bevy.workspace = true
-bevy_mod_picking = { version = "0.20", features = ["backend_bevy_ui"] }
+bevy_mod_picking = { workspace = true, optional = true }
 rand.workspace = true
 metrics-exporter-prometheus = {workspace = true, optional = true}
 bevy-inspector-egui.workspace = true
